feat: extract the createObservers method
