+++ packages/ui/src/components/DataStory/clients/ClientBuffer.ts
@@ -1,17 +1,19 @@
 import { InputObserver, ItemValue } from '@data-story/core';
 import { bufferTime, concatMap, Observable, pipe, UnaryFunction } from 'rxjs';
 
-function groupNotifications(data: {
-  items: ItemValue[],
-  inputObservers: InputObserver[]
-}[]) {
+interface ObserverParams {
+  items: ItemValue[];
+  inputObservers: InputObserver[];
+}
+
+function groupNotifications(data: ObserverParams[]) {
   const getNotifyObserverKey = (items: ItemValue[], inputObservers: InputObserver[]) => {
     return inputObservers.map((inputObserver) => {
       return inputObserver.observerId;
     }).sort().join(',');
   }
 
-  const groupedNotificationMap = new Map<string, {items: ItemValue[], inputObservers: InputObserver[]}>();
+  const groupedNotificationMap = new Map<string, ObserverParams>();
   data.forEach((d) => {
     const key = getNotifyObserverKey(d.items, d.inputObservers);
     if (groupedNotificationMap.has(key)) {
@@ -29,10 +31,7 @@ function groupNotifications(data: {
   return Array.from(groupedNotificationMap.values());
 }
 
-export function clientBuffer(bufferTimeSpan: number = 500): UnaryFunction<Observable<{
-  items: ItemValue[],
-  inputObservers: InputObserver[]
-}>, Observable<{items: ItemValue[], inputObservers: InputObserver[]}>> {
+export function clientBuffer(bufferTimeSpan: number = 500): UnaryFunction<Observable<ObserverParams>, Observable<ObserverParams>> {
   return pipe(
     bufferTime(bufferTimeSpan),
     concatMap(groupNotifications)
